requirement.txt

langchain 
langchain_groq 
fastapi 
uvicorn 
pymongo 
python-dotenv